requirements.txt
discord.py
logging
asyncio
typing
sys
humanfriendly
datetime